courses_backend/requirements.txt
Django==5.1
djangorestframework==3.14.0
django-cors-headers==4.2.0
